LayoutTests/platform/ios/fast/visual-viewport/ios/caret-after-focus-in-fixed-expected.txt

Focusing first field
Moving to next field
FAIL finalCaretRect.left.toFixed(2) should be 44.00. Was 43.00.
PASS finalCaretRect.top.toFixed(2) is "1244.00"
PASS successfullyParsed is true

TEST COMPLETE
Checks the location of the caret after focussing inputs inside position:fixed.
